Update README.md
When I tried to convert the md file to pdf, it did not work for me in the command line (GitBash)
Also, although I created a file titled "pandoc_transformations.zip." However, when I uploaded the files into my repository it displayed as 3 separate files.
This was a great exercise! Thank you! :)